+++ CHANGELOG.md
@@ -4,12 +4,17 @@ For updates follow [@implydata](https://twitter.com/implydata) on Twitter.
 
 ## 0.16.1
 
+**Major breaking changes in this release**
+
 - Allow count distinct on cross product (e.g. `$data.countDistinct($a ++ $b)`)
 - Transition to TS 2.1
 - Better division by 0 handling
 - Extracted Requester type definition into `plywood-base-api`
 - Switched from `Q` to `any-promise`
 - Use has-own-prop library
+- Changed `Dataset#toJS` to return an object rather than just the `data` array
+- Datasets now actively maintain their `keys` array
+- `Dataset#flatten` now returns a new (flat) `Dataset` and thus the `parentName` option is no longer supported
 
 
 ## 0.15.11